SONARJAVA-3663 Deprecate rule S2976 in favor of S5445
