+++ java-checks/src/main/resources/org/sonar/l10n/java/rules/java/Sonar_way_profile.json
@@ -255,7 +255,6 @@
     "S2925",
     "S2970",
     "S2975",
-    "S2976",
     "S3008",
     "S3010",
     "S3011",
@@ -380,6 +379,7 @@
     "S5411",
     "S5413",
     "S5443",
+    "S5445",
     "S5527",
     "S5542",
     "S5547",
